core: tsc sync on host write
Updated handle_msr_write to assume that any host-initiated write to
IA32_TSC with a value of 0 means that the host is attempting to
synchronize the TSC offsets across cpus, and thus that the same offset
value should be written into all vcpus.

In situations where the host's TSCs are in-sync, this should result in
in-sync guest TSCs. When the guest TSCs are in-sync, the TSC clocksource
in the linux guest should be more reliable and monotonic.

Bug: 159829614
Test: Ran a tsc synchronization test script, and ran a VM successfully
  with the TSC clocksource
Change-Id: I917b90af6b9a3a7b8ab4fbc8c6c04507312ff61a